changed double to float

+++ cadi_main_d.c
@@ -16,7 +16,7 @@
 
 #define PI 3.14159265358979323846
 
-void transpose(float* restrict u_t, const float* restrict u, const int n){
+void transpose(double* restrict u_t, const double* restrict u, const int n){
     for (int i = 0; i < n; i++) {
         for (int j = 0; j < n; j++) {
             u_t[j*n+i] = u[i*n+j];
@@ -24,9 +24,9 @@ void transpose(float* restrict u_t, const float* restrict u, const int n){
     }
 }
 
-//void apply_tri(float* restrict vOut, const float* restrict vIn,
-//               const float* restrict diag,  const float* restrict udiag,
-//               const float* restrict ldiag, const int n){
+//void apply_tri(double* restrict vOut, const double* restrict vIn,
+//               const double* restrict diag,  const double* restrict udiag,
+//               const double* restrict ldiag, const int n){
 //    vOut[0] = diag[0] * vIn[0] + udiag[0] * vIn[1];
 //    vOut[n-1] = diag[n-1] * vIn[n-1] + ldiag[n-2] * vIn[n-2];
 //    for(int i = 1; i < n-1; i++){
@@ -34,16 +34,16 @@ void transpose(float* restrict u_t, const float* restrict u, const int n){
 //    }
 //}
 
-//void solve_tri(float* restrict x, const float* restrict ldiag,
-//               const float* restrict diag, const float* restrict udiag,
+//void solve_tri(double* restrict x, const double* restrict ldiag,
+//               const double* restrict diag, const double* restrict udiag,
 //               const int n) {
 //    // Allocate scratch space.
-//    float* cprime = (float*) malloc(sizeof(float) * n);
+//    double* cprime = (double*) malloc(sizeof(double) * n);
 //    cprime[0] = udiag[0] / diag[0];
 //    x[0] = x[0] / diag[0];
 //    // loop from 1 to N - 1 inclusive
 //    for (int i = 1; i < n; i++) {
-//        float m = 1.0 / (diag[i] - ldiag[i-1] * cprime[i - 1]);
+//        double m = 1.0 / (diag[i] - ldiag[i-1] * cprime[i - 1]);
 //        cprime[i] = udiag[i] * m;
 //        x[i] = (x[i] - ldiag[i-1] * x[i - 1]) * m;
 //    }
@@ -54,22 +54,22 @@ void transpose(float* restrict u_t, const float* restrict u, const int n){
 //    free(cprime);
 //}
 
-void apply_tri_special(float* restrict u, float* restrict u_temp, const float a,  const float b, const int n){
-    /*
+void apply_tri_special(double* restrict u, double* restrict u_temp, const double a,  const double b, const int n){
+    /* 
      This operator is designed to apply a tridiagonal matrix multiplication with the following
      structure:
-
+     
      0  0   0   0   0   0
      a  b   a   0   0   0
      0  a   b   a   0   0
      ......
      0  0   0   a   b   a
      0  0   0   0   0   0
-
+     
      There is little room for generalizing its application but it should yield great performance boost
      */
-//    float * u_copy = (float*) malloc(n * sizeof(float));
-    memcpy(u_temp, u, n*sizeof(float));
+//    double * u_copy = (double*) malloc(n * sizeof(double));
+    memcpy(u_temp, u, n*sizeof(double));
     for(int i = 1; i < n-1; i++){
         u[i] = a*(u_temp[i-1]+u_temp[i+1]) + b * u_temp[i];
     }
@@ -77,49 +77,49 @@ void apply_tri_special(float* restrict u, float* restrict u_temp, const float a,
     u[n-1] = 0.0;
 //    free(u_copy);
 }
-void apply_tri_special_plus(float* restrict u, float* restrict u_temp, const float a,  const float b, const int n){
+void apply_tri_special_plus(double* restrict u, double* restrict u_temp, const double a,  const double b, const int n){
     /*
      This operator is designed to apply a tridiagonal matrix multiplication with the following
      structure:
-
+     
      1  0   0   0   0   0
      a  b+1 a   0   0   0
      0  a   b+1 a   0   0
      ......
      0  0   0   a   b+1 a
      0  0   0   0   0   1
-
+     
      There is little room for generalizing its application but it should yield great performance boost
      */
-//    float* scratch = (float*) malloc(sizeof(float) * n);
-    memcpy(u_temp, u, n*sizeof(float));
+//    double* scratch = (double*) malloc(sizeof(double) * n);
+    memcpy(u_temp, u, n*sizeof(double));
     for(int i = 1; i < n-1; i++){
         u[i] += a*(u_temp[i-1]+u_temp[i+1]) + b * u_temp[i];
     }
 //    free(scratch);
 }
 
-void solve_tri_special(float* restrict x, const float a, const float b, const int n){
-//void solve_tri_special(float* restrict x, float* restrict cprime, const float a, const float b, const int n){
-     /*
+void solve_tri_special(double* restrict x, const double a, const double b, const int n){
+//void solve_tri_special(double* restrict x, double* restrict cprime, const double a, const double b, const int n){
+     /* 
      This solve is designed to solve a tridiagonal linear system using thomas' algorithm with the following structure:
-
+     
      1  0   0   0   0   0
      -a 1-b -a  0   0   0
      0  -a  1-b -a  0   0
      ......
      0  0   0   -a  1-b -a
      0  0   0   0   0   1
-
+     
      There is little room for generalizing its application but it should yield great performance boost
      */
-
+    
     // Allocate scratch space.
-    float* restrict cprime __attribute__((aligned(64))) = (float*) _mm_malloc(sizeof(float) * n, 64);
+    double* restrict cprime __attribute__((aligned(64))) = (double*) _mm_malloc(sizeof(double) * n, 64);
     cprime[0] = 0;
-    float astar = -a;
-    float bstar = 1-b;
-    float m;
+    double astar = -a;
+    double bstar = 1-b;
+    double m;
     // loop from 1 to N - 2 inclusive
     for (int i = 1; i < n-1; i++) {
         m = 1.0 / (bstar - astar * cprime[i - 1]);
@@ -130,15 +130,15 @@ void solve_tri_special(float* restrict x, const float a, const float b, const in
     // loop from N - 2 to 0 inclusive, safely testing loop end condition
     for (int i = n - 1; i-- > 0; )
         x[i] = x[i] - cprime[i] * x[i + 1];
-
+    
     _mm_free(cprime);
 }
 
-void relaxOperation(float * restrict u, const float * restrict fstar, float* scratch, float a, float b, int n){
-
-    float* ustar = (float*) malloc(n*n*sizeof(float));
-    float* u_t = (float*) malloc(n*n*sizeof(float));
-    memcpy(ustar, u, n*n*sizeof(float));
+void relaxOperation(double * restrict u, const double * restrict fstar, double* scratch, double a, double b, int n){
+    
+    double* ustar = (double*) malloc(n*n*sizeof(double));
+    double* u_t = (double*) malloc(n*n*sizeof(double));
+    memcpy(ustar, u, n*n*sizeof(double));
     transpose(u_t, ustar, n);
     // keep two copies and then do it.
     for (int i = 1; i < n-1; i++) {
@@ -147,8 +147,8 @@ void relaxOperation(float * restrict u, const float * restrict fstar, float* scr
     for (int i = 1; i < n-1; i++) {
         apply_tri_special(u_t+i*n, scratch, a, b, n);
     }
-    memset(u_t, 0.0, n*sizeof(float));
-    memset(u_t+(n-1)*n, 0.0, n*sizeof(float));
+    memset(u_t, 0.0, n*sizeof(double));
+    memset(u_t+(n-1)*n, 0.0, n*sizeof(double));
     for (int i = 0; i < n*n; i++) {
         u[i] -= fstar[i];
     }
@@ -175,20 +175,20 @@ void relaxOperation(float * restrict u, const float * restrict fstar, float* scr
         solve_tri_special(ustar+i*n, a/2, b/2, n);
     }
     transpose(u, ustar, n);
-
+    
     free(ustar);
     free(u_t);
 }
-float normInf(const float* restrict u, const int n){
-    float max = u[0];
+double normInf(const double* restrict u, const int n){
+    double max = u[0];
     for (int i = 1; i < n*n; i++) {
         max = (max < fabs(u[i]))? fabs(u[i]):max;
     }
     return max;
 }
-float test_init(const float x, const float y, const float wx, const float wy, const float ax, const float ay){
-    float d2Xfact = (2.0*PI*wx)*(2.0*PI*wx);
-    float d2Yfact = (2.0*PI*wy)*(2.0*PI*wy);
+double test_init(const double x, const double y, const double wx, const double wy, const double ax, const double ay){
+    double d2Xfact = (2.0*PI*wx)*(2.0*PI*wx);
+    double d2Yfact = (2.0*PI*wy)*(2.0*PI*wy);
     return -(cos(2.0*PI*x*wx)*cos(2.0*PI*y*wy))/(ax*d2Xfact + ay*d2Yfact);
 }
 
@@ -197,7 +197,7 @@ float test_init(const float x, const float y, const float wx, const float wy, co
 int main(int argc, char * argv[]) {
     int M = 200; // Problem size, assume square case
     int threadCount = -1;
-
+    
     int c;
     extern char* optarg;
     const char* optstring = "n:p:";
@@ -207,38 +207,38 @@ int main(int argc, char * argv[]) {
             case 'p': threadCount = atoi(optarg); break;
         }
     }
-
-    float alpha       = 2.0;
-    float waveNumber  = 1.0;
-    float alphaX      = alpha;     // Laplace operator-x prefactor
-    float alphaY      = alpha;     // Laplace operator-y prefactor
-    float waveNumberX = waveNumber;  // test problem x-coordinate wave number
-    float waveNumberY = waveNumber;  // test problem y-coordinate wave number
-
+    
+    double alpha       = 2.0;
+    double waveNumber  = 1.0;
+    double alphaX      = alpha;     // Laplace operator-x prefactor
+    double alphaY      = alpha;     // Laplace operator-y prefactor
+    double waveNumberX = waveNumber;  // test problem x-coordinate wave number
+    double waveNumberY = waveNumber;  // test problem y-coordinate wave number
+    
     int   xPanel = M;  // X panel count
     int   yPanel = M;  // Y panel count
-
-    float xMin = 0.0; float xMax = 1.0;
-    float yMin = 0.0; float yMax = 1.0;
-    float hx = (xMax-xMin)/(float)xPanel;
-    float hy = (yMax-yMin)/(float)yPanel;
-
-    float dt_0 = 4*hx*hx/(alpha*PI*PI);  // Relaxation timestep
-
+    
+    double xMin = 0.0; double xMax = 1.0;
+    double yMin = 0.0; double yMax = 1.0;
+    double hx = (xMax-xMin)/(double)xPanel;
+    double hy = (yMax-yMin)/(double)yPanel;
+    
+    double dt_0 = 4*hx*hx/(alpha*PI*PI);  // Relaxation timestep
+    
     // Echo input parameters
     printf("\n==== Douglas ADI Program Start ====\n");
     printf("** alpha_x/alpha_y : %f/%f\n", alphaX, alphaY);
     printf("** X/Y Panel Count : %d/%d\n", xPanel, yPanel);
-
+    
     int N = M+1; // vector size
-    //float* f      = (float*) malloc(N*N*sizeof(float));
-    //float* uk     = (float*) calloc(N*N, sizeof(float)); // initialize to 0
-    //float* uLast  = (float*) malloc(N*N*sizeof(float));
-    float* restrict f __attribute__((aligned(64))) = (float*) _mm_malloc(N*N*sizeof(float),64);
-    float* restrict uk __attribute__((aligned(64))) = (float*) _mm_malloc(N*N*sizeof(float),64);
-    float* restrict uLast __attribute__((aligned(64))) = (float*) _mm_malloc(N*N*sizeof(float),64);
+    //double* f      = (double*) malloc(N*N*sizeof(double));
+    //double* uk     = (double*) calloc(N*N, sizeof(double)); // initialize to 0
+    //double* uLast  = (double*) malloc(N*N*sizeof(double));
+    double* restrict f __attribute__((aligned(64))) = (double*) _mm_malloc(N*N*sizeof(double),64);
+    double* restrict uk __attribute__((aligned(64))) = (double*) _mm_malloc(N*N*sizeof(double),64);
+    double* restrict uLast __attribute__((aligned(64))) = (double*) _mm_malloc(N*N*sizeof(double),64);
     memset(uk, 0.0, N*N);
-    float tol = 1.0e-6;  // Stopping tolerance
+    double tol = 1.0e-6;  // Stopping tolerance
 
 
     __assume_aligned(f, 64);
@@ -253,14 +253,14 @@ int main(int argc, char * argv[]) {
      |
      | (i)
      |
-
+     
      */
-    float x;
-    float y;
+    double x;
+    double y;
 //    for (int i = 0; i < N*N; i++) {
 //        uk[i] = i;
 //    }
-
+    
     for(int j = 0; j < N; j++){
         y =  yMin + j*hy;
         for(int i = 0; i < N; i++){
@@ -296,38 +296,38 @@ int main(int argc, char * argv[]) {
         uk[index + i*N]  = test_init(x, y, waveNumberX, waveNumberY, alphaX, alphaY);
         f [index + i*N]  = 0.0;
     }
-
+    
    // Set up OpenMP parameters
 #ifdef _OPENMP
     if(threadCount > omp_get_max_threads() || threadCount <= 0) omp_set_num_threads(omp_get_max_threads());
     else omp_set_num_threads(threadCount);
     printf("\n==>Using OpenMP With %d Threads\n\n",omp_get_max_threads());
 #endif
-
+    
     // Initialize solver variables
     const int maxSweepSize = (int)ceil(log(M)/log(2)) + 1;
-    float dt = dt_0;
+    double dt = dt_0;
 
-    float ax = dt*alphaX/(hx*hx);
-    float ay = dt*alphaY/(hy*hy);
-    float bx = -2.0 * ax;
-    float by = -2.0 * ay;
-
-    float diffNorm = 2*tol;
+    double ax = dt*alphaX/(hx*hx);
+    double ay = dt*alphaY/(hy*hy);
+    double bx = -2.0 * ax;
+    double by = -2.0 * ay;
+    
+    double diffNorm = 2*tol;
     int   iterMax  = 200;
     int   iter     = 0;
-    float ukNorm = 0;
-
+    double ukNorm = 0;
+    
 //    ax = 1;
 //    bx = 2;
-//    float a[5] = {4,18,3,1,1};
-//    float b[5];
+//    double a[5] = {4,18,3,1,1};
+//    double b[5];
 //    solve_tri_special(a, ax, bx, 5);
 //    apply_tri_special_plus(a, ax, bx, 5);
 //    b[0]=b[0];
-
-    //float* fstar = (float*) malloc(maxSweepSize*N*N*sizeof(float));
-    float* restrict fstar __attribute__((aligned(64))) = (float*) _mm_malloc(maxSweepSize*N*N*sizeof(float),64);
+    
+    //double* fstar = (double*) malloc(maxSweepSize*N*N*sizeof(double));
+    double* restrict fstar __attribute__((aligned(64))) = (double*) _mm_malloc(maxSweepSize*N*N*sizeof(double),64);
     for (int i = 0; i < maxSweepSize; i++) {
         int offset = i*N*N;
         for (int j = 0; j < N*N; j++) {
@@ -335,11 +335,11 @@ int main(int argc, char * argv[]) {
         }
         dt *= 2*2;
     }
-
-    float* restrict scratch __attribute__((aligned(64))) = (float*) _mm_malloc(N*sizeof(float),64);
-
+    
+    double* restrict scratch __attribute__((aligned(64))) = (double*) _mm_malloc(N*sizeof(double),64);
+    
     printf("\n");
-
+    
 //    relaxOperation(uk, fstar, scratch, ax, bx, N);
 //    for (int i = 0 ; i < N; i++) {
 //        for (int j = 0; j < N; j++) {
@@ -348,9 +348,9 @@ int main(int argc, char * argv[]) {
 //        printf("\n");
 //    }
 
-    float t0 = omp_get_wtime();
+    double t0 = omp_get_wtime();
     while((diffNorm > tol)&&(iter < iterMax)){
-        memcpy(uLast, uk, N*N*sizeof(float)); // This can be avoided by swapping the roles for each iteration
+        memcpy(uLast, uk, N*N*sizeof(double)); // This can be avoided by swapping the roles for each iteration
         for (int j = 0; j < maxSweepSize; j++) {
             relaxOperation(uk, fstar+j*N*N, scratch, ax, bx, N);
 //            dt *= 2*2;
@@ -364,7 +364,7 @@ int main(int argc, char * argv[]) {
 //            }
 //            printf("\n");
 //        }
-//
+//        
         for (int j = maxSweepSize-1; j > -1 ; j--) {
 //            dt /= 2*2;
             ax /= 2*2;
@@ -378,14 +378,14 @@ int main(int argc, char * argv[]) {
         diffNorm = normInf(uLast, N) / ukNorm; // uLast.normInf();
         iter++; // Update iterate
     }
-    float t1 = omp_get_wtime();
-
+    double t1 = omp_get_wtime();
+    
     _mm_free(scratch);
     _mm_free(fstar);
     _mm_free(f);
     _mm_free(uk);
     _mm_free(uLast);
-
+    
     printf("==== Multi-scale Timstep Relaxation Output XXXX ====\n");
     printf("** Difference between iterates : \t\t\t\t%3.10f\n",diffNorm);
     printf("** Iteration Count for multi-scale time scheme :\t %d \n\n",iter*2*maxSweepSize);
@@ -393,6 +393,6 @@ int main(int argc, char * argv[]) {
 //    printf("** Time it takes to initialize the multi-scale timestep solution is %e ms\n", elapsed_1);
     printf("** Time it takes to solve the multi-scale timestep solution is %e ms\n", t1-t0);
 //    printf("** The total time is is %e ms\n", elapsed_1+elapsed_2);
-
+    
     return 0;
 }
